+++ ff-core/lib/FF/CLI.hs
@@ -24,7 +24,9 @@ import Data.Text.Prettyprint.Doc
     PageWidth (AvailablePerLine),
     defaultLayoutOptions,
     layoutPageWidth,
-    layoutSmart
+    layoutSmart,
+    pretty,
+    vsep
     )
 import Data.Text.Prettyprint.Doc.Render.Terminal (AnsiStyle, renderStrict)
 import Data.Time (Day)
@@ -48,6 +50,7 @@ import FF
     getWikiSamples,
     loadAllTagTexts,
     noDataDirectoryMessage,
+    sponsors,
     updateTrackedNotes
     )
 import FF.Config
@@ -189,6 +192,7 @@ runCmdAction ui cmd isBrief = do
     CmdTags -> do
       allTags <- loadAllTagTexts
       pprint $ prettyTagsList allTags
+    CmdSponsors -> pprint $ withHeader "Sponsors" $ vsep $ map pretty sponsors
     CmdTrack track ->
       cmdTrack track today isBrief
     CmdUnarchive tasks ->