+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to
 
 ## [Unreleased]
 ### Added
-- Nix build.
-- Tags.
+- Nix build
+- Sponsor list under subcommand `ff sponsors`
+- Tags
 
 ### Changed
-- Updated dependencies.
+- Updated dependencies
+- Set all fields to empty when deleting note or contact instead of assigning
+  status `Deleted`
 
 ### Fixed
 - GTK build.